requirements.txt
Flask
gunicorn
mysql-connector-python
pandas
openpyxl
reportlab
